[Dev-24] Upgraded Menu Bar
GameFrame now has a "Settings" menu.
The "Settings" menu has menu items to set the game parameters through #setGameParameters.

GameFrame now has a "Help" menu.
The "Help" menu has items to acquire a window with a game tutorial and license information.

The menu bar also features a New Game button, which is performs an idempotent call for a new game.